requirements.txt
asgiref==3.6.0
Django==4.2
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
asgiref==3.6.0
astroid==2.15.4
black==23.3.0
click==8.1.3
dill==0.3.6
Django==4.2
djangorestframework==3.14.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.3.0
pylint==2.17.3
pylint-django==2.5.3
pylint-plugin-utils==0.7
pytz==2023.3
sqlparse==0.4.4
tomlkit==0.11.7
wrapt==1.15.0
